Abstracted out the OpenGL window example
Abstracted out the deafult functionality of opening a window to a 'Window' class with a specific platform specific class (in this case 'WindowsWindow'). I also created an 'Application' class for handling the application state, such as creating and managing a window and updating the application behaviours through 'Application Layers'.

Also created a simple 'Log' class for logging default text, warnings and errors and implemented debugging macros for debugging internal errors.

Finally I created an 'Event' class that has all the basic events in it. These events are then invoked through the OpenGL window event callbacks.